send invite emails/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.10.0
google-cloud-secret-manager==2.10.0